fix(dynamic-tags): address second review wave
- helpers: designsetgo_resolve_bindings_post_id honours a pre-set
  __dsgo_post_id in \$args so Registry::resolve() (invoked from the
  REST preview with no WP_Block) no longer falls through to
  get_the_ID() and return null for every scalar source
- field-discovery: skip ACF 'password' field type so plaintext
  password fields never surface as a bindable source in the picker
- extensions/bindable-attributes: wrap per-attribute labels in __()
  so they're translatable
- sources-post/site: move the shared image sub-value helper into
  ImageResolver::resolve_subvalue; shrinks sources-post.php under
  the 300-line guideline
- components/DynamicTagPicker: split out SourceSidebar, SourceArgsForm,
  and PreviewPanel into their own files; main component now 181 lines
