+++ includes/blocks/dynamic-tags/class-dynamic-tags-sources-post.php
@@ -227,7 +227,7 @@ class PostSources {
 				}
 				$subkey = isset( $args['subkey'] ) ? (string) $args['subkey'] : 'url';
 				$size   = isset( $args['size'] ) ? (string) $args['size'] : 'full';
-				return self::resolve_image_subvalue( $attachment_id, $subkey, $size );
+				return ImageResolver::resolve_subvalue( $attachment_id, $subkey, $size );
 			},
 			array(
 				'subkey' => array(
@@ -296,49 +296,5 @@ class PostSources {
 				'args'    => $args_schema,
 			)
 		);
-	}
-
-	/**
-	 * Resolves a scalar sub-value from an attachment ID.
-	 *
-	 * @param int    $attachment_id Attachment post ID.
-	 * @param string $subkey        Sub-key: url|id|alt|width|height|title|caption.
-	 * @param string $size          Image size for url/width/height.
-	 * @return string|null
-	 */
-	public static function resolve_image_subvalue( $attachment_id, $subkey, $size = 'full' ) {
-		$attachment_id = (int) $attachment_id;
-		if ( ! $attachment_id ) {
-			return null;
-		}
-		if ( ! in_array( $subkey, array( 'url', 'id', 'alt', 'width', 'height', 'title', 'caption' ), true ) ) {
-			$subkey = 'url';
-		}
-
-		switch ( $subkey ) {
-			case 'id':
-				return (string) $attachment_id;
-			case 'alt':
-				$alt = get_post_meta( $attachment_id, '_wp_attachment_image_alt', true );
-				return '' === $alt ? null : (string) $alt;
-			case 'title':
-				$title = get_the_title( $attachment_id );
-				return '' === $title ? null : (string) $title;
-			case 'caption':
-				$caption = wp_get_attachment_caption( $attachment_id );
-				return '' === $caption || false === $caption ? null : (string) $caption;
-			case 'width':
-			case 'height':
-				$src = wp_get_attachment_image_src( $attachment_id, $size );
-				if ( ! $src ) {
-					return null;
-				}
-				$idx = 'width' === $subkey ? 1 : 2;
-				return (string) $src[ $idx ];
-			case 'url':
-			default:
-				$src = wp_get_attachment_image_src( $attachment_id, $size );
-				return $src && ! empty( $src[0] ) ? (string) $src[0] : null;
-		}
 	}
 }
