+++ src/components/DynamicTagPicker/DynamicTagPicker.js
@@ -19,21 +19,19 @@ import { useState, useEffect, useMemo } from '@wordpress/element';
 import {
 	Modal,
 	Button,
-	SelectControl,
-	TextControl,
-	SearchControl,
 	__experimentalVStack as VStack,
 	__experimentalHStack as HStack,
-	Spinner,
-	Notice,
 } from '@wordpress/components';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
 
 import { useDynamicTagSources } from './useDynamicTagSources';
 import { useDynamicTagFields } from './useDynamicTagFields';
 import { useDynamicTagPreview } from './useDynamicTagPreview';
+import SourceSidebar from './SourceSidebar';
+import SourceArgsForm from './SourceArgsForm';
+import PreviewPanel from './PreviewPanel';
 import './style.scss';
 
 const DEFAULT_RETURNS = [ 'text', 'url', 'image', 'number', 'date' ];
@@ -63,7 +61,7 @@ export default function DynamicTagPicker( {
 	const postId = postIdProp || editorContext.postId;
 	const postType = postTypeProp || editorContext.postType || 'post';
 
-	const { status: sourcesStatus, groups, sources, error: sourcesError } = useDynamicTagSources( { returns } );
+	const { status: sourcesStatus, groups, sources } = useDynamicTagSources( { returns } );
 
 	const [ selectedSource, setSelectedSource ] = useState( value?.source || '' );
 	const [ draftArgs, setDraftArgs ] = useState( value?.args || {} );
@@ -96,24 +94,6 @@ export default function DynamicTagPicker( {
 		size: draftArgs?.size,
 	} );
 
-	const filteredSources = useMemo( () => {
-		if ( ! search ) {
-			return sources;
-		}
-		const needle = search.toLowerCase();
-		return sources.filter( ( s ) => s.label.toLowerCase().includes( needle ) || s.slug.toLowerCase().includes( needle ) );
-	}, [ sources, search ] );
-
-	const groupedSources = useMemo( () => {
-		const bucket = {};
-		filteredSources.forEach( ( s ) => {
-			( bucket[ s.group ] = bucket[ s.group ] || [] ).push( s );
-		} );
-		return groups
-			.filter( ( g ) => bucket[ g.slug ]?.length )
-			.map( ( g ) => ( { ...g, sources: bucket[ g.slug ] } ) );
-	}, [ filteredSources, groups ] );
-
 	if ( ! isOpen ) {
 		return null;
 	}
@@ -122,16 +102,18 @@ export default function DynamicTagPicker( {
 		if ( ! selectedSource ) {
 			return;
 		}
-		onChange( {
-			source: selectedSource,
-			args: draftArgs,
-		} );
+		onChange( { source: selectedSource, args: draftArgs } );
 		onClose?.();
 	};
 
 	const handleClear = () => {
 		onChange( null );
 		onClose?.();
+	};
+
+	const handleSelectSource = ( slug ) => {
+		setSelectedSource( slug );
+		setDraftArgs( {} );
 	};
 
 	return (
@@ -142,52 +124,15 @@ export default function DynamicTagPicker( {
 			size="large"
 		>
 			<div className="dsgo-dynamic-tag-picker__layout">
-				<div className="dsgo-dynamic-tag-picker__sidebar">
-					<SearchControl
-						value={ search }
-						onChange={ setSearch }
-						placeholder={ __( 'Search sources…', 'designsetgo' ) }
-						__nextHasNoMarginBottom
-					/>
-
-					{ sourcesStatus === 'loading' && (
-						<div className="dsgo-dynamic-tag-picker__loading"><Spinner /></div>
-					) }
-
-					{ sourcesStatus === 'error' && (
-						<Notice status="error" isDismissible={ false }>
-							{ __( 'Unable to load Dynamic Tag sources.', 'designsetgo' ) }
-						</Notice>
-					) }
-
-					{ sourcesStatus === 'ready' && groupedSources.length === 0 && (
-						<p className="dsgo-dynamic-tag-picker__empty">
-							{ __( 'No sources match.', 'designsetgo' ) }
-						</p>
-					) }
-
-					{ groupedSources.map( ( group ) => (
-						<div key={ group.slug } className="dsgo-dynamic-tag-picker__group">
-							<h3 className="dsgo-dynamic-tag-picker__group-title">{ group.label }</h3>
-							<ul className="dsgo-dynamic-tag-picker__source-list">
-								{ group.sources.map( ( source ) => (
-									<li key={ source.slug }>
-										<Button
-											variant={ source.slug === selectedSource ? 'primary' : 'tertiary' }
-											onClick={ () => {
-												setSelectedSource( source.slug );
-												setDraftArgs( {} );
-											} }
-											className="dsgo-dynamic-tag-picker__source-button"
-										>
-											{ source.label }
-										</Button>
-									</li>
-								) ) }
-							</ul>
-						</div>
-					) ) }
-				</div>
+				<SourceSidebar
+					status={ sourcesStatus }
+					groups={ groups }
+					sources={ sources }
+					search={ search }
+					onSearchChange={ setSearch }
+					selectedSource={ selectedSource }
+					onSelectSource={ handleSelectSource }
+				/>
 
 				<div className="dsgo-dynamic-tag-picker__detail">
 					{ ! activeSource && (
@@ -232,170 +177,5 @@ export default function DynamicTagPicker( {
 				</Button>
 			</HStack>
 		</Modal>
-	);
-}
-
-function SourceArgsForm( { source, args, onChange, fieldDiscovery } ) {
-	const schema = source.args || {};
-	const entries = Object.entries( schema );
-
-	if ( entries.length === 0 ) {
-		return null;
-	}
-
-	const setArg = ( key, value ) => {
-		const next = { ...args };
-		if ( value === '' || value === undefined || value === null ) {
-			delete next[ key ];
-		} else {
-			next[ key ] = value;
-		}
-		onChange( next );
-	};
-
-	return (
-		<VStack spacing={ 3 }>
-			{ entries.map( ( [ argName, argSchema ] ) => {
-				// ACF/meta "key" field — surface discovered fields as a select when available.
-				if ( argName === 'key' && source.supportsFieldDiscovery ) {
-					const fieldOptions = [
-						{ label: __( '— Select a field —', 'designsetgo' ), value: '' },
-						...fieldDiscovery.fields.map( ( f ) => ( {
-							label: f.group ? `${ f.group } — ${ f.label }` : f.label,
-							value: f.key,
-						} ) ),
-					];
-					return (
-						<div key={ argName }>
-							{ fieldDiscovery.status === 'loading' ? (
-								<Spinner />
-							) : (
-								<SelectControl
-									label={ __( 'Field', 'designsetgo' ) }
-									value={ args[ argName ] || '' }
-									options={ fieldOptions }
-									onChange={ ( value ) => setArg( argName, value ) }
-									__nextHasNoMarginBottom
-									__next40pxDefaultSize
-								/>
-							) }
-							<TextControl
-								label={ __( 'Or enter a field key manually', 'designsetgo' ) }
-								value={ args[ argName ] || '' }
-								onChange={ ( value ) => setArg( argName, value ) }
-								__nextHasNoMarginBottom
-								__next40pxDefaultSize
-							/>
-						</div>
-					);
-				}
-
-				if ( Array.isArray( argSchema.enum ) && argSchema.enum.length > 0 ) {
-					return (
-						<SelectControl
-							key={ argName }
-							label={ argLabel( argName ) }
-							value={ args[ argName ] ?? argSchema.default ?? '' }
-							options={ [
-								{ label: __( '— Default —', 'designsetgo' ), value: '' },
-								...argSchema.enum.map( ( v ) => ( { label: v, value: v } ) ),
-							] }
-							onChange={ ( value ) => setArg( argName, value ) }
-							__nextHasNoMarginBottom
-							__next40pxDefaultSize
-						/>
-					);
-				}
-
-				return (
-					<TextControl
-						key={ argName }
-						label={ argLabel( argName ) }
-						help={ argSchema.description || '' }
-						value={ args[ argName ] ?? '' }
-						onChange={ ( value ) => setArg( argName, value ) }
-						__nextHasNoMarginBottom
-						__next40pxDefaultSize
-					/>
-				);
-			} ) }
-		</VStack>
-	);
-}
-
-function argLabel( name ) {
-	const map = {
-		key: __( 'Field key', 'designsetgo' ),
-		subkey: __( 'Sub-value', 'designsetgo' ),
-		size: __( 'Image size', 'designsetgo' ),
-		scope: __( 'Scope', 'designsetgo' ),
-		format: __( 'Date format', 'designsetgo' ),
-		taxonomy: __( 'Taxonomy', 'designsetgo' ),
-		separator: __( 'Separator', 'designsetgo' ),
-	};
-	return map[ name ] || name;
-}
-
-function PreviewPanel( { preview, returns = [] } ) {
-	const isImage = returns.includes( 'image' );
-
-	if ( preview.status === 'idle' ) {
-		return null;
-	}
-
-	if ( preview.status === 'loading' ) {
-		return (
-			<div className="dsgo-dynamic-tag-picker__preview">
-				<Spinner />
-			</div>
-		);
-	}
-
-	if ( preview.status === 'empty' ) {
-		return (
-			<Notice status="info" isDismissible={ false }>
-				{ __( 'Preview is empty for the current post.', 'designsetgo' ) }
-			</Notice>
-		);
-	}
-
-	if ( preview.status === 'unauthorized' ) {
-		return (
-			<Notice status="warning" isDismissible={ false }>
-				{ __( 'Preview is hidden because the post is password-protected or private.', 'designsetgo' ) }
-			</Notice>
-		);
-	}
-
-	if ( preview.status !== 'resolved' ) {
-		return (
-			<Notice status="error" isDismissible={ false }>
-				{ __( 'Unable to preview this source.', 'designsetgo' ) }
-			</Notice>
-		);
-	}
-
-	if ( isImage && preview.value && typeof preview.value === 'object' ) {
-		return (
-			<div className="dsgo-dynamic-tag-picker__preview">
-				<h4>{ __( 'Preview', 'designsetgo' ) }</h4>
-				<img src={ preview.value.url } alt={ preview.value.alt || '' } />
-				<p className="dsgo-dynamic-tag-picker__preview-meta">
-					{ sprintf(
-						/* translators: %1$s image width, %2$s image height */
-						__( '%1$s × %2$s', 'designsetgo' ),
-						preview.value.width || '?',
-						preview.value.height || '?'
-					) }
-				</p>
-			</div>
-		);
-	}
-
-	return (
-		<div className="dsgo-dynamic-tag-picker__preview">
-			<h4>{ __( 'Preview', 'designsetgo' ) }</h4>
-			<p className="dsgo-dynamic-tag-picker__preview-value">{ String( preview.value ) }</p>
-		</div>
 	);
 }
